+++ spring-cloud-gcp-security-iap/src/main/java/org/springframework/cloud/gcp/security/iap/IapAuthenticationFilter.java
@@ -21,6 +21,7 @@ import javax.servlet.http.HttpServletRequest;
 import org.apache.commons.logging.Log;
 import org.apache.commons.logging.LogFactory;
 
+import org.springframework.cloud.gcp.security.iap.jwt.JwtTokenVerifier;
 import org.springframework.security.core.Authentication;
 import org.springframework.security.web.authentication.preauth.AbstractPreAuthenticatedProcessingFilter;
